requirements.txt
# 데이터 처리 및 모델 학습을 위해 필요한 패키지 목록
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.6.1
lightgbm==4.5.0
catboost==1.2.7
xgboost==2.1.4
matplotlib==3.10.0
seaborn==0.13.2
scipy==1.13.1
pyyaml==6.0.1